botme-robot-ui/requirements.txt
Pillow~=8.3.2
colour~=0.1.5
PyAudio~=0.2.11
websocket-client==1.2.1
gTTS~=2.2.3
pip~=21.2.4
click~=8.0.1
six~=1.16.0
urllib3~=1.26.6
requests~=2.26.0
gevent~=21.8.0
greenlet~=1.1.1
certifi~=2021.5.30
idna~=3.2
setuptools~=53.0.0
cachetools==4.2.2
charset-normalizer==2.0.4
google-api-core==2.0.1
google-auth==2.0.2
google-cloud-speech==2.8.0
googleapis-common-protos==1.53.0
grpcio==1.40.0
libcst==0.3.20
mypy-extensions==0.4.3
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyYAML==5.4.1
rsa==4.7.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
zope.event==4.5.0
zope.interface==5.4.0
python-socketio==5.5.0
